Reformat motd module and factor out ERB instantiation method
