+++ apps/dashboard/app/controllers/dashboard_controller.rb
@@ -1,10 +1,6 @@
 class DashboardController < ApplicationController
   def index
-    begin
-      @motd = MotdFile.new.formatter
-    rescue StandardError => e
-      flash.now[:alert] = t('dashboard.motd_erb_render_error', error_message: e.message)
-    end
+    @motd = MotdFile.new.formatter
     set_my_quotas
   end
