--- apps/dashboard/app/models/motd_formatter/markdown.rb
@@ -1,17 +1,0 @@
-module MotdFormatter
-  class Markdown
-
-    attr_reader :content, :title
-
-    # @param [MotdFile] motd_file an MotdFile object that contains a URI path to a message of the day in OSC format
-    def initialize(motd_file)
-      motd_file = MotdFile.new unless motd_file
-      @title = motd_file.title
-      @content = OodAppkit.markdown.render(motd_file.content)
-    end
-
-    def to_partial_path
-      "dashboard/motd_markdown"
-    end
-  end
-end
